refactor(factory): extract SuggestedFiles, StagedColumn, ShippedColumn sub-components
DetailPanel.svelte was 523 lines and FactoryFloor.svelte was 567 lines (S1 violations).
Extracted:
- SuggestedFiles.svelte: semantic suggested-files list with styles
- StagedColumn.svelte: Kommissionierung column with release controls
- ShippedColumn.svelte: Versand column with PR links

Results: DetailPanel → 473 lines, FactoryFloor → 499 lines (both under 500 limit).
