+++ website/src/components/factory/ShippedColumn.svelte
@@ -1,0 +1,58 @@
+<script lang="ts">
+  let {
+    shipped,
+    mobileColIndex,
+    onOpenDetail,
+    relTime,
+    prUrl,
+    ticketUrl,
+  }: {
+    shipped: { extId: string; title: string; prNumber?: number | null; doneAt?: string | null }[];
+    mobileColIndex: number;
+    onOpenDetail: (extId: string) => void;
+    relTime: (iso: string | null) => string;
+    prUrl: (n: number) => string;
+    ticketUrl: (extId: string) => string;
+  } = $props();
+</script>
+
+<div data-col="done" class:mobile-visible={mobileColIndex === 9} class="lg:w-1/5" data-testid="floor-shipped">
+  <h3 class="font-semibold mb-2">Versand</h3>
+  {#if shipped.length === 0}
+    <p class="text-muted text-sm">Noch nichts versandt.</p>
+  {:else}
+    <ul class="space-y-1.5">
+      {#each shipped as s (s.extId)}
+        <li class="rounded-lg border border-transparent bg-white/5 px-2.5 py-2 text-sm transition-colors hover:border-white/10 hover:bg-white/[0.08]"
+            data-testid="floor-shipped-item">
+          <div class="flex items-center justify-between gap-2">
+            <a href={ticketUrl(s.extId)} class="font-mono text-xs text-gold hover:underline"
+               title="In der Ticket-Übersicht öffnen">{s.extId}</a>
+            {#if s.doneAt}
+              <span class="whitespace-nowrap text-[10px] text-muted"
+                    title={new Date(s.doneAt).toLocaleString('de-DE')}>{relTime(s.doneAt)}</span>
+            {/if}
+          </div>
+          <button type="button" onclick={() => onOpenDetail(s.extId)}
+                  class="mt-0.5 block w-full text-left leading-snug transition-colors hover:text-gold"
+                  title="Phasen-Timeline &amp; Details anzeigen">{s.title}</button>
+          {#if s.prNumber}
+            <a href={prUrl(s.prNumber)} target="_blank" rel="noopener noreferrer"
+               data-testid="floor-shipped-pr"
+               class="mt-1 inline-flex items-center gap-1 rounded bg-white/5 px-1.5 py-0.5 text-[11px] font-medium transition-colors hover:bg-gold hover:text-dark">
+              <svg viewBox="0 0 16 16" class="h-3 w-3" fill="currentColor" aria-hidden="true"><path d="M1.5 3.25a2.25 2.25 0 1 1 3 2.122v5.256a2.251 2.251 0 1 1-1.5 0V5.372A2.25 2.25 0 0 1 1.5 3.25Zm5.677-.177L9.573.677A.25.25 0 0 1 10 .854V2.5h1A2.5 2.5 0 0 1 13.5 5v5.628a2.251 2.251 0 1 1-1.5 0V5a1 1 0 0 0-1-1h-1v1.646a.25.25 0 0 1-.427.177L7.177 3.427a.25.25 0 0 1 0-.354ZM3.75 2.5a.75.75 0 1 0 0 1.5.75.75 0 0 0 0-1.5Zm0 9.5a.75.75 0 1 0 0 1.5.75.75 0 0 0 0-1.5Zm8.25.75a.75.75 0 1 0 1.5 0 .75.75 0 0 0-1.5 0Z"/></svg>
+              PR #{s.prNumber}<span class="opacity-60">↗</span>
+            </a>
+          {/if}
+        </li>
+      {/each}
+    </ul>
+  {/if}
+</div>
+
+<style>
+  @media (max-width: 767px) {
+    [data-col] { display: none; }
+    [data-col].mobile-visible { display: flex; flex-direction: column; width: 100%; }
+  }
+</style>
